+++ scripts/src/netlist.lua
@@ -62,6 +62,7 @@ project "netlist"
 		MAME_DIR .. "src/lib/netlist/plib/pdynlib.h",
 		MAME_DIR .. "src/lib/netlist/plib/pmain.cpp",
 		MAME_DIR .. "src/lib/netlist/plib/pmain.h",
+    MAME_DIR .. "src/lib/netlist/plib/pmath.h",
 		MAME_DIR .. "src/lib/netlist/plib/pmempool.h",
 		MAME_DIR .. "src/lib/netlist/plib/pomp.h",
 		MAME_DIR .. "src/lib/netlist/plib/poptions.cpp",
@@ -69,7 +70,8 @@ project "netlist"
 		MAME_DIR .. "src/lib/netlist/plib/ppmf.h",
     MAME_DIR .. "src/lib/netlist/plib/ppreprocessor.cpp",
     MAME_DIR .. "src/lib/netlist/plib/ppreprocessor.h",
-		MAME_DIR .. "src/lib/netlist/plib/pstate.h",
+    MAME_DIR .. "src/lib/netlist/plib/pstate.h",
+		MAME_DIR .. "src/lib/netlist/plib/pstonum.h",
 		MAME_DIR .. "src/lib/netlist/plib/pstring.cpp",
 		MAME_DIR .. "src/lib/netlist/plib/pstring.h",
 		MAME_DIR .. "src/lib/netlist/plib/pstrutil.h",
