+++ src/lib/netlist/analog/nlid_twoterm.h
@@ -39,8 +39,6 @@
 #include "nld_generic_models.h"
 #include "plib/pfunction.h"
 
-#include <cmath>
-
 // -----------------------------------------------------------------------------
 // Implementation
 // -----------------------------------------------------------------------------
