+++ src/lib/netlist/devices/nld_74123.cpp
@@ -8,8 +8,6 @@
 #include "nlid_system.h"
 #include "netlist/analog/nlid_twoterm.h"
 
-#include <cmath>
-
 namespace netlist
 {
 	namespace devices
