+++ src/lib/netlist/nl_base.cpp
@@ -19,8 +19,6 @@
 
 #include "nl_errstr.h"
 
-#include <cmath>
-//#include <cstring>
 #include <limits>
 
 namespace netlist
