+++ src/lib/netlist/plib/pfmtlog.cpp
@@ -8,6 +8,7 @@
 #include "pfmtlog.h"
 #include "palloc.h"
 #include "pstrutil.h"
+#include "pstonum.h"
 
 #include <algorithm>
 #include <array>
