+++ src/lib/netlist/plib/ppreprocessor.cpp
@@ -8,6 +8,7 @@
 #include "ppreprocessor.h"
 #include "palloc.h"
 #include "putil.h"
+#include "pstonum.h"
 
 namespace plib {
