+++ src/lib/netlist/plib/ptypes.h
@@ -92,7 +92,6 @@ namespace plib
 		return !(v & (v-1));
 	}
 
-
 	//============================================================
 	// abs, lcd, gcm
 	//============================================================
