+++ src/lib/netlist/plib/vector_ops.h
@@ -14,7 +14,6 @@
 
 #include <algorithm>
 #include <array>
-#include <cmath>
 #include <type_traits>
 
 #if !defined(__clang__) && !defined(_MSC_VER) && (__GNUC__ > 4 || (__GNUC__ == 4 && __GNUC_MINOR__ > 6))
@@ -137,7 +136,7 @@ namespace plib
 	{
 		T ret = 0.0;
 		for ( std::size_t i = 0; i < n; i++ )
-			ret = std::max(ret, std::abs(v[i]));
+			ret = std::max(ret, plib::abs(v[i]));
 
 		return ret;
 	}
