+++ src/lib/netlist/solver/nld_matrix_solver.cpp
@@ -8,8 +8,6 @@
 #include "nld_matrix_solver.h"
 #include "plib/putil.h"
 
-#include <cmath>  // <<= needed by windows build
-
 namespace netlist
 {
 namespace solver
