+++ src/lib/netlist/solver/nld_ms_direct.h
@@ -14,7 +14,6 @@
 #include "plib/vector_ops.h"
 
 #include <algorithm>
-#include <cmath>
 
 namespace netlist
 {
