+++ src/lib/netlist/solver/nld_solver.cpp
@@ -46,7 +46,6 @@
 #include "plib/pomp.h"
 
 #include <algorithm>
-#include <cmath>
 
 namespace netlist
 {
